refactor, conda fixes (#851)
* Renamed peoplegroup to usergroup, removed yubikey

* kill jupyter notebook

* Biobanking and consent stuff gone (#837)

* Refactor auditing system (#839)

* more cleanup

* Fix tests

* umask support to UI

* merge

* rename roles -> services, services -> groups

* fixes

* update deploy script

* fixes

* rename RoleAudit to ServiceAudit

* updates

* more fixes

* more fixes

* finished refactoring, i think

* first-time login - fill in username/pwd, warn pwd needs changing

* first time login

* first time login

* host_id, not hostid

* conda search fix, no leaking of kagent passwd

* fix jupyter_settings

* fix

* destroy anaconda support

* order conda_commands by id before returning to kagent

* fixes

* fix alerts to have a FK on hosts

* jupyter_settings

* fixes

* conda fixes

* cleanup ui

* button for syncing anaconda in admin/hosts

* enable tf_serving as part of the tensorflow tour

* remove jupyter notebook rows  correctly from jupyter_project

* add gpus to hosts

* fix jupyter pid to be 0 if orphaned

* more support for anaconda synchronization for new nodes. hosts table now has a has_gpus column

* improve msg for anaconda preparation

* put distro info into parent pom for maven deployments

* fixes to show python version, admin_password, cleanup jupyter kernels when removing a conda env

* change to snapshot version

* num_gpus, not has_gpu

* fix deploy script

* deploy script fix

* fix

* antonios fix for zeppelin

+++ hopsworks-common/src/main/java/io/hops/hopsworks/common/dao/jupyter/config/JupyterFacade.java
@@ -12,7 +12,6 @@ import io.hops.hopsworks.common.exception.AppException;
 import io.hops.hopsworks.common.hdfs.HdfsUsersController;
 import io.hops.hopsworks.common.util.Settings;
 import java.io.File;
-import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 import java.util.logging.Level;
@@ -30,7 +29,7 @@ import javax.ws.rs.core.Response;
 public class JupyterFacade {
 
   private static final Logger logger = Logger.getLogger(JupyterFacade.class.
-          getName());
+      getName());
 
   @PersistenceContext(unitName = "kthfsPU")
   private EntityManager em;
@@ -52,29 +51,38 @@ public class JupyterFacade {
     return em;
   }
 
-  public List<JupyterProject> findNotebooksByProject(Integer projectId) {
-    TypedQuery<JupyterProject> query = em.createNamedQuery(
-            "JupyterProject.findByProjectId",
-            JupyterProject.class);
-    query.setParameter("projectId", projectId);
-    List<JupyterProject> res = query.getResultList();
-    List<JupyterProject> notebooks = new ArrayList<>();
-    for (JupyterProject pt : res) {
-//      notebooks.add(new TopicDTO(pt.getProjectTopicsPK().getTopicName(),
-//              pt.getSchemaTopics().getSchemaTopicsPK().getName(),
-//              pt.getSchemaTopics().getSchemaTopicsPK().getVersion()));
-    }
-    return notebooks;
-  }
-
+//  public List<JupyterProject> findNotebooksByProject(Integer projectId) {
+//    TypedQuery<JupyterProject> query = em.createNamedQuery(
+//        "JupyterProject.findByProjectId",
+//        JupyterProject.class);
+//    query.setParameter("projectId", projectId);
+//    List<JupyterProject> res = query.getResultList();
+//    List<JupyterProject> notebooks = new ArrayList<>();
+//    for (JupyterProject pt : res) {
+////      notebooks.add(new TopicDTO(pt.getProjectTopicsPK().getTopicName(),
+////              pt.getSchemaTopics().getSchemaTopicsPK().getName(),
+////              pt.getSchemaTopics().getSchemaTopicsPK().getVersion()));
+//    }
+//    return notebooks;
+//  }
   public boolean removeNotebookServer(String hdfsUsername) {
+
+    if (hdfsUsername == null || hdfsUsername.isEmpty()) {
+      return false;
+    }
 
     JupyterProject jp = findByUser(hdfsUsername);
     if (jp == null) {
       return false;
     }
-    em.remove(jp);
-    em.flush();
+    try {
+      em.remove(jp);
+      em.flush();
+    } catch (Exception ex) {
+      logger.warning("Problem removing jupyter notebook entry from hopsworks DB");
+      logger.warning(ex.getMessage());
+      return false;
+    }
     return true;
   }
 
@@ -109,24 +117,24 @@ public class JupyterFacade {
   public JupyterProject findByUser(String hdfsUser) {
     HdfsUsers res = null;
     TypedQuery<HdfsUsers> query = em.createNamedQuery(
-            "HdfsUsers.findByName", HdfsUsers.class);
+        "HdfsUsers.findByName", HdfsUsers.class);
     query.setParameter("name", hdfsUser);
     try {
       res = query.getSingleResult();
     } catch (EntityNotFoundException | NoResultException e) {
       Logger.getLogger(JupyterFacade.class.getName()).log(Level.FINE, null,
-              e);
+          e);
       return null;
     }
     JupyterProject res2 = null;
     TypedQuery<JupyterProject> query2 = em.createNamedQuery(
-            "JupyterProject.findByHdfsUserId", JupyterProject.class);
+        "JupyterProject.findByHdfsUserId", JupyterProject.class);
     query2.setParameter("hdfsUserId", res.getId());
     try {
       res2 = query2.getSingleResult();
     } catch (EntityNotFoundException | NoResultException e) {
       Logger.getLogger(JupyterFacade.class.getName()).log(Level.FINE, null,
-              e);
+          e);
     }
     return res2;
   }
@@ -135,7 +143,7 @@ public class JupyterFacade {
 
     if (hdfsUser == null) {
       throw new AppException(Response.Status.NOT_FOUND.getStatusCode(),
-              "Could not find a Jupyter Notebook server to delete.");
+          "Could not find a Jupyter Notebook server to delete.");
     }
 
     JupyterProject jp = this.findByUser(hdfsUser);
@@ -145,12 +153,12 @@ public class JupyterFacade {
   public List<JupyterProject> getAllNotebookServers() {
     List<JupyterProject> res = null;
     TypedQuery<JupyterProject> query = em.createNamedQuery(
-            "JupyterProject.findAll", JupyterProject.class);
+        "JupyterProject.findAll", JupyterProject.class);
     try {
       res = query.getResultList();
     } catch (EntityNotFoundException | NoResultException e) {
       Logger.getLogger(JupyterFacade.class.getName()).log(Level.FINE, null,
-              e);
+          e);
       return null;
     }
     return res;
@@ -162,14 +170,14 @@ public class JupyterFacade {
   }
 
   public JupyterProject saveServer(String host,
-          Project project, String secretConfig, int port,
-          int hdfsUserId, String token, long pid)
-          throws AppException {
+      Project project, String secretConfig, int port,
+      int hdfsUserId, String token, long pid)
+      throws AppException {
     JupyterProject jp = null;
     String ip;
     ip = host + ":" + settings.getHopsworksPort();
     jp = new JupyterProject(project, secretConfig, port, hdfsUserId, ip, token,
-            pid);
+        pid);
 
     persist(jp);
     return jp;
@@ -211,9 +219,9 @@ public class JupyterFacade {
   }
 
   public String getProjectPath(JupyterProject jp, String projectName,
-          String hdfsUser) {
+      String hdfsUser) {
     return settings.getJupyterDir() + File.separator
-            + Settings.DIR_ROOT + File.separator + projectName
-            + File.separator + hdfsUser + File.separator + jp.getSecret();
+        + Settings.DIR_ROOT + File.separator + projectName
+        + File.separator + hdfsUser + File.separator + jp.getSecret();
   }
 }
